requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
pymongo==4.6.0
python-dotenv==1.0.0
